refactor: renaming files and element tag
