+++ packages/xlayers/src/app/core/sketch.service.mock.ts
@@ -1,5 +1,5 @@
 import { SketchData } from './sketch.service';
-import { getFlatLayerMock } from '@app/editor/preview/viewer/sketch-layer/sketch-layer.component.mock';
+import { getFlatLayerMock } from '@app/editor/preview/viewer/layer/layer.component.mock';
 
 export function getSketchDataMock() {
   return {
