+++ packages/xlayers/src/app/editor/preview/viewer/page/page.component.spec.ts
@@ -1,10 +1,10 @@
 import { async, ComponentFixture, TestBed } from '@angular/core/testing';
 
-import { SketchPageComponent } from './sketch-page.component';
+import { SketchPageComponent } from './page.component';
 import { NO_ERRORS_SCHEMA } from '@angular/core';
 import { NgxsModule } from '@ngxs/store';
 import { By } from '@angular/platform-browser';
-import { getFrameMock } from '../sketch-layer/sketch-layer.component.mock';
+import { getFrameMock } from '../layer/layer.component.mock';
 import { UiState } from '@app/core/state';
 import { CodeGenState } from '@app/core/state/page.state';
 import { MatSnackBarModule } from '@angular/material/snack-bar';
