+++ packages/xlayers/src/app/editor/preview/viewer/stop-event-propagation.directive.ts
