+++ packages/xlayers/src/app/editor/preview/viewer/viewer.module.ts
@@ -1,13 +1,13 @@
 import { NgModule } from '@angular/core';
 import { CoreModule } from '@app/core/core.module';
 import { SketchParserModule } from '@xlayers/sketchapp-parser';
-import { SketchCanvasComponent } from './sketch-canvas/sketch-canvas.component';
-import { SketchContainerComponent } from './sketch-container/sketch-container.component';
-import { SketchSelectedLayerDirective } from './sketch-layer/selected-layer.directive';
-import { SketchLayerComponent } from './sketch-layer/sketch-layer.component';
-import { SketchPageComponent } from './sketch-page/sketch-page.component';
-import { SketchStopEventPropagationDirective } from './sketch-stop-event-propagation.directive';
-import { SketchTooltilDirective } from './sketch-tooltip.directive';
+import { SketchCanvasComponent } from './canvas/canvas.component';
+import { SketchContainerComponent } from './container/container.component';
+import { SketchSelectedLayerDirective } from './layer/selected-layer.directive';
+import { SketchLayerComponent } from './layer/layer.component';
+import { SketchPageComponent } from './page/page.component';
+import { SketchStopEventPropagationDirective } from './stop-event-propagation.directive';
+import { SketchTooltilDirective } from './tooltip.directive';
 
 @NgModule({
   imports: [CoreModule, SketchParserModule],
