+++ packages/xlayers/src/app/upload/select-demo-files/select-demo-files.component.html
@@ -29,6 +29,6 @@
     </ng-container>
   </mat-nav-list>
 </mat-expansion-panel>
-<p class="sketch-select-demo-files-label-error" *ngIf="error">
+<p class="select-demo-files-label-error" *ngIf="error">
   Error: please choose different file
 </p>
