+++ packages/xlayers/src/app/upload/upload.component.html
@@ -4,9 +4,9 @@
   <span class="mat-subheading-1">OR</span>
 
   <xly-browse-files (changed)="onFileSelected($event)"></xly-browse-files>
-  
-  <p class="sketch-select-demo-files-label">OR</p>
-  
+
+  <p class="select-demo-files-label">OR</p>
+
   <xly-select-demo-files
     [error]="selectedDemoFileError"
     (changed)="openSelectedDemoFile($event)"
